# Conflicts:
#	source/YaneuraOu.vcxproj
#	source/YaneuraOu.vcxproj.filters

+++ source/usi.cpp
@@ -251,14 +251,14 @@ namespace USI
 				auto pos_ = const_cast<Position*>(&pos);
 				Move moves[MAX_PLY + 1];
 				StateInfo si[MAX_PLY];
-				int ply = 0;
+				size_t ply = 0;
 
 				while ( ply < MAX_PLY )
 				{
 					// 千日手はそこで終了。ただし初手はPVを出力。
 					// 千日手がベストのとき、置換表を更新していないので
 					// 置換表上はMOVE_NONEがベストの指し手になっている可能性があるので早めに検出する。
-					auto rep = pos.is_repetition(ply);
+					auto rep = pos.is_repetition(int(ply));
 					if (rep != REPETITION_NONE && ply >= 1)
 					{
 						// 千日手でPVを打ち切るときはその旨を表示
